legacytx: move RegisterLegacyAminoCodec from init to tests/init
