+++ x/auth/legacy/legacytx/codec.go
@@ -4,7 +4,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec"
 )
 
-func init() {
-	var amino = codec.NewLegacyAmino()
-	amino.RegisterConcrete(StdTx{}, "cosmos-sdk/StdTx", nil)
+func RegisterLegacyAminoCodec(cdc *codec.LegacyAmino) {
+	cdc.RegisterConcrete(StdTx{}, "cosmos-sdk/StdTx", nil)
 }
