+++ x/auth/legacy/legacytx/stdtx.go
@@ -21,9 +21,10 @@ var (
 	_ sdk.FeeTx          = (*StdTx)(nil)
 )
 
-// Deprecated: StdFee includes the amount of coins paid in fees and the maximum
+// StdFee includes the amount of coins paid in fees and the maximum
 // gas to be used by the transaction. The ratio yields an effective "gasprice",
 // which must be above some miminum to be accepted into the mempool.
+// [Deprecated]
 type StdFee struct {
 	Amount sdk.Coins `json:"amount" yaml:"amount"`
 	Gas    uint64    `json:"gas" yaml:"gas"`
